requirements_test.txt
-r requirements.txt
pytest-homeassistant-custom-component==0.1.0